# auth_service/docs/swagger.yaml
basePath: /
definitions:
...
  internal_http_server_handlers_account_request_restore_confirmation.Request:
    properties:
      app_id:
        example: 1
        type: integer
      email:
        example: example@domain.com
        type: string
    required:
    - app_id
    - email
    type: object
  internal_http_server_handlers_account_restore.Request:
    properties:
      email:
        example: example@domain.com
        type: string
      password:
        example: SecurePass123!
        type: string
      session_id:
        example: fkajeDJ1p3FJ...
        type: string
      token:
        example: abcDEF123...
        type: string
    required:
    - email
    type: object
...
host: localhost:8082
info:
  contact: {}
  description: Сервис авторизации
  title: Auth Service API
  version: "1.0"
paths:
...
  /account/restore:
    post:
      consumes:
      - application/json
      description: |-
        Отменяет soft-delete, если grace period (7 дней) ещё не
        истёк. Требует подтверждения: паролем (если он установлен)
        либо magic-link кодом, полученным через
        /account/restore/request-confirmation (для oauth-only
        пользователей без пароля). Неаутентифицированный эндпоинт.
      parameters:
      - description: Email + (пароль ИЛИ session_id+code)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_http_server_handlers_account_restore.Request'
      produces:
      - application/json
      responses:
        "204":
          description: Аккаунт восстановлен
        "400":
          description: Невалидный запрос
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
        "401":
          description: Неверный пароль или код подтверждения, аккаунт не найден, не
            был удалён или grace period истёк
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
        "429":
          description: Превышен лимит запросов
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
      summary: Восстановить удалённый аккаунт
      tags:
      - account
  /account/restore/request-confirmation:
    post:
      consumes:
      - application/json
      description: |-
        Отправляет magic-link код на email указанного (soft-deleted)
        аккаунта для подтверждения восстановления. Неаутентифицированный
        эндпоинт — юзер не может залогиниться, пока аккаунт удалён.
        Возвращает session_id для последующего запроса в /account/restore.
      parameters:
      - description: Email и app_id
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_http_server_handlers_account_request_restore_confirmation.Request'
      produces:
      - application/json
      responses:
        "200":
          description: Код отправлен на email
          schema:
            properties:
              session_id:
                type: string
              status:
                type: string
            type: object
        "400":
          description: Невалидный запрос
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
        "429":
          description: Превышен лимит запросов
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
      summary: Запросить подтверждение восстановления аккаунта через magic link
      tags:
      - account
...
  /auth/2fa/magic-link/verify:
    post:
      consumes:
      - application/json
      description: |-
        Завершает второй фактор аутентификации: проверяет токен из
        письма в связке с session_id, полученным на этапе /auth/login,
        и при успехе выдаёт access/refresh токены. Токен одноразовый —
        повторное использование того же токена или session_id отклоняется.
      parameters:
      - description: Данные для подтверждения
        in: body
        name: request
        required: true
        schema:
          properties:
            session_id:
              type: string
            token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 2FA подтверждена, выданы токены
          schema:
            properties:
              access_token:
                type: string
              refresh_token:
                type: string
              status:
                type: string
            type: object
        "400":
          description: Невалидное тело запроса или ошибка валидации
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
        "401":
          description: Токен невалиден, истёк, уже использован, либо сессия истекла
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
      summary: Подтверждение magic-link 2FA
      tags:
      - 2fa
...
swagger: "2.0"